Added check in ljl_loggedin() for if ljl_getljuser() returns the magical "?UNKNOWN!" pseudo-user for failed uid map lookup.
Also pushed the removal of cookies at logout into its own function, so
that it can trash all the relevant cookies in a session at once, and
also put it in a try/except, 'cause anything that uses a component
should really do that.

Finally, I need to start making one set of changes to a file at a time.
Having more than one cause for a set of edits is naughty of me.