Fixed the name of the class for when there's no username/uid map match.
Also, not that it shows in this edit, because I undid it before checkin,
but I'm grumpy that the function to pull an entry from the PM throws an
exception if there's no match instead of, oh, I don't know, returning
false or just not populating the result fields or something sane. Which
means I'm going to have to do iterator crap, which I was hoping to
avoid. Cockbites.