YES! FINALLY! Last of the bugs squashed, and now 1.2.0 is FEATURE-COMPLETE! HALLELUJAH!
Oh. The last bug? If you had a default login set, and it didn't
complete, either due to problems or because you hit "Cancel" when
prompted for the password for a non-PM-saved default, you got the green
"LJlogin" instead of a "Logged out". So, put in a call to
ljl_loggedout() first so that's set no matter what, to be overridden if
a login happens.